fix: validação de linha não selecionada

+++ modulo-rh/src/main/java/br/com/ifba/modulorh/pagamentos/view/TelaDeListarPagamentos.java
@@ -453,6 +453,11 @@ public class TelaDeListarPagamentos extends javax.swing.JFrame {
         try{
             Pagamentos pagamento = new Pagamentos();
             int linha = tblPagamentos.getSelectedRow();
+             if (linha == -1){
+                JOptionPane.showMessageDialog(null, "Nenhuma linha da tabela selecionada! ",
+                "Nenhuma linha da tabela selecionada!",JOptionPane.ERROR_MESSAGE);
+                return ;
+            }
             Long id = (Long) tblPagamentos.getValueAt(linha, 0);
             pagamento.setId(id);
             facade.deletePagamento(pagamento);
